backend/app/requirements.txt
paho-mqtt
pandas
aiohttp
pyserial
websockets
